Add a BaseSchema class and move methods that are purely duplicated down.
